moving lots of files, removing unused stuff
