+++ macos/Sources/Features/Dock Tile Plugin/Notification+AppIcon.swift
@@ -1,0 +1,5 @@
+import AppKit
+
+extension Notification.Name {
+    static let ghosttyIconDidChange = Notification.Name("com.mitchellh.ghostty.iconDidChange")
+}
